+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fixed LICENSE generation to use BitMinded as copyright holder instead of product name
+- Updated existing repository licenses (rythmo, measure-mate) to use BitMinded proprietary license
+
 ---
 
 ## [1.0.3] - 2025-01-29